refactor: arrange direcotry and module
