+++ Cargo.toml
@@ -4,8 +4,14 @@ version = "0.4.0"
 authors = ["forestaa <t.daichi.3@gmail.com>"]
 edition = "2018"
 
+[lib]
+name = "troll"
+doctest = false
+path = "src/lib.rs"
+
 [[bin]]
 name = "troll"
+test = false
 path = "src/main.rs"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
