+++ src/domain/mod.rs
@@ -1,5 +1,6 @@
 pub mod global_variable;
 pub mod global_variable_view;
+pub mod global_variable_view_factory;
 pub mod global_variables_extractor;
 pub mod type_entry;
 pub mod type_entry_repository;
