+++ src/lib.rs
@@ -1,0 +1,4 @@
+pub mod domain;
+pub mod infrastructure;
+pub mod library;
+pub mod usecase;
