+++ src/library/dwarf.rs
@@ -381,615 +381,115 @@ impl IntoIterator for DwarfInfoIntoIterator {
     }
 }
 
-#[cfg(test)]
-pub mod tests {
-    use super::*;
+pub struct DwarfInfoBuilder<OffsetP, TagP> {
+    offset: OffsetP,
+    tag: TagP,
+    name: Option<String>,
+    type_offset: Option<Offset>,
+    byte_size: Option<usize>,
+    location: Option<Location>,
+    upper_bound: Option<usize>,
+    data_member_location: Option<usize>,
+    children: Vec<DwarfInfo>,
+}
 
-    pub struct DwarfInfoBuilder<OffsetP, TagP> {
-        offset: OffsetP,
-        tag: TagP,
-        name: Option<String>,
-        type_offset: Option<Offset>,
-        byte_size: Option<usize>,
-        location: Option<Location>,
-        upper_bound: Option<usize>,
-        data_member_location: Option<usize>,
-        children: Vec<DwarfInfo>,
-    }
-
-    impl DwarfInfoBuilder<(), ()> {
-        pub fn new() -> Self {
-            DwarfInfoBuilder {
-                offset: (),
-                tag: (),
-                name: None,
-                type_offset: None,
-                byte_size: None,
-                location: None,
-                upper_bound: None,
-                data_member_location: None,
-                children: Vec::new(),
-            }
+impl DwarfInfoBuilder<(), ()> {
+    pub fn new() -> Self {
+        DwarfInfoBuilder {
+            offset: (),
+            tag: (),
+            name: None,
+            type_offset: None,
+            byte_size: None,
+            location: None,
+            upper_bound: None,
+            data_member_location: None,
+            children: Vec::new(),
         }
     }
+}
 
-    impl DwarfInfoBuilder<Offset, DwarfTag> {
-        pub fn build(self) -> DwarfInfo {
-            DwarfInfo {
-                offset: self.offset,
-                tag: self.tag,
-                name: self.name,
-                type_offset: self.type_offset,
-                byte_size: self.byte_size,
-                location: self.location,
-                upper_bound: self.upper_bound,
-                data_member_location: self.data_member_location,
-                children: self.children,
-            }
+impl DwarfInfoBuilder<Offset, DwarfTag> {
+    pub fn build(self) -> DwarfInfo {
+        DwarfInfo {
+            offset: self.offset,
+            tag: self.tag,
+            name: self.name,
+            type_offset: self.type_offset,
+            byte_size: self.byte_size,
+            location: self.location,
+            upper_bound: self.upper_bound,
+            data_member_location: self.data_member_location,
+            children: self.children,
         }
     }
+}
 
-    impl<OffsetP> DwarfInfoBuilder<OffsetP, ()> {
-        pub fn tag(self, tag: DwarfTag) -> DwarfInfoBuilder<OffsetP, DwarfTag> {
-            DwarfInfoBuilder {
-                offset: self.offset,
-                tag: tag,
-                name: self.name,
-                type_offset: self.type_offset,
-                byte_size: self.byte_size,
-                location: self.location,
-                upper_bound: self.upper_bound,
-                data_member_location: self.data_member_location,
-                children: self.children,
-            }
+impl<OffsetP> DwarfInfoBuilder<OffsetP, ()> {
+    pub fn tag(self, tag: DwarfTag) -> DwarfInfoBuilder<OffsetP, DwarfTag> {
+        DwarfInfoBuilder {
+            offset: self.offset,
+            tag: tag,
+            name: self.name,
+            type_offset: self.type_offset,
+            byte_size: self.byte_size,
+            location: self.location,
+            upper_bound: self.upper_bound,
+            data_member_location: self.data_member_location,
+            children: self.children,
         }
     }
+}
 
-    impl<TagP> DwarfInfoBuilder<(), TagP> {
-        pub fn offset(self, offset: Offset) -> DwarfInfoBuilder<Offset, TagP> {
-            DwarfInfoBuilder {
-                offset: offset,
-                tag: self.tag,
-                name: self.name,
-                type_offset: self.type_offset,
-                byte_size: self.byte_size,
-                location: self.location,
-                upper_bound: self.upper_bound,
-                data_member_location: self.data_member_location,
-                children: self.children,
-            }
+impl<TagP> DwarfInfoBuilder<(), TagP> {
+    pub fn offset(self, offset: Offset) -> DwarfInfoBuilder<Offset, TagP> {
+        DwarfInfoBuilder {
+            offset: offset,
+            tag: self.tag,
+            name: self.name,
+            type_offset: self.type_offset,
+            byte_size: self.byte_size,
+            location: self.location,
+            upper_bound: self.upper_bound,
+            data_member_location: self.data_member_location,
+            children: self.children,
         }
     }
+}
 
-    impl<OffsetP, TagP> DwarfInfoBuilder<OffsetP, TagP> {
-        pub fn name<S: Into<String>>(mut self, name: S) -> Self {
-            self.name = Some(name.into());
-            self
-        }
-
-        pub fn type_offset(mut self, type_offset: Offset) -> Self {
-            self.type_offset = Some(type_offset);
-            self
-        }
-
-        pub fn byte_size(mut self, size: usize) -> Self {
-            self.byte_size = Some(size);
-            self
-        }
-
-        pub fn location(mut self, location: Location) -> Self {
-            self.location = Some(location);
-            self
-        }
-
-        pub fn upper_bound(mut self, upper_bound: usize) -> Self {
-            self.upper_bound = Some(upper_bound);
-            self
-        }
-
-        pub fn data_member_location(mut self, data_member_location: usize) -> Self {
-            self.data_member_location = Some(data_member_location);
-            self
-        }
-
-        pub fn children(mut self, children: Vec<DwarfInfo>) -> Self {
-            self.children = children;
-            self
-        }
+impl<OffsetP, TagP> DwarfInfoBuilder<OffsetP, TagP> {
+    pub fn name<S: Into<String>>(mut self, name: S) -> Self {
+        self.name = Some(name.into());
+        self
     }
 
-    fn init() {
-        let _ = env_logger::builder().is_test(true).try_init();
+    pub fn type_offset(mut self, type_offset: Offset) -> Self {
+        self.type_offset = Some(type_offset);
+        self
     }
 
-    fn dwarf_info_intoiterator_test<S: Into<String>>(elf_path: S, expected: Vec<DwarfInfo>) {
-        init();
-
-        let got: Vec<DwarfInfo> = DwarfInfoIntoIterator::new(elf_path.into())
-            .into_iter()
-            .collect();
-        assert_eq!(expected, got);
+    pub fn byte_size(mut self, size: usize) -> Self {
+        self.byte_size = Some(size);
+        self
     }
 
-    #[test]
-    #[ignore]
-    fn dwarf_info_const() {
-        let expected = vec![
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(45))
-                .tag(DwarfTag::DW_TAG_variable)
-                .name("c")
-                .type_offset(Offset::new(72))
-                .location(Location::new(8196))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(65))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(4)
-                .name("int")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(72))
-                .tag(DwarfTag::DW_TAG_const_type)
-                .type_offset(Offset::new(65))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(77))
-                .tag(DwarfTag::DW_TAG_unimplemented)
-                .name("main")
-                .type_offset(Offset::new(65))
-                .build(),
-        ];
-
-        dwarf_info_intoiterator_test("examples/const", expected);
+    pub fn location(mut self, location: Location) -> Self {
+        self.location = Some(location);
+        self
     }
 
-    #[test]
-    #[ignore]
-    fn dwarf_info_pointer() {
-        let expected = vec![
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(45))
-                .tag(DwarfTag::DW_TAG_variable)
-                .name("p")
-                .type_offset(Offset::new(65))
-                .location(Location::new(16432))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(65))
-                .tag(DwarfTag::DW_TAG_pointer_type)
-                .byte_size(8)
-                .type_offset(Offset::new(71))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(71))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(4)
-                .name("int")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(78))
-                .tag(DwarfTag::DW_TAG_unimplemented)
-                .name("main")
-                .type_offset(Offset::new(71))
-                .build(),
-        ];
-
-        dwarf_info_intoiterator_test("examples/pointer", expected);
+    pub fn upper_bound(mut self, upper_bound: usize) -> Self {
+        self.upper_bound = Some(upper_bound);
+        self
     }
 
-    #[test]
-    #[ignore]
-    fn dwarf_info_typedef() {
-        let expected = vec![
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(45))
-                .tag(DwarfTag::DW_TAG_typedef)
-                .name("uint8")
-                .type_offset(Offset::new(57))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(57))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(4)
-                .name("unsigned int")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(64))
-                .tag(DwarfTag::DW_TAG_variable)
-                .name("a")
-                .type_offset(Offset::new(45))
-                .location(Location::new(16428))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(84))
-                .tag(DwarfTag::DW_TAG_unimplemented)
-                .name("main")
-                .type_offset(Offset::new(114))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(114))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(4)
-                .name("int")
-                .build(),
-        ];
-
-        dwarf_info_intoiterator_test("examples/typedef", expected);
+    pub fn data_member_location(mut self, data_member_location: usize) -> Self {
+        self.data_member_location = Some(data_member_location);
+        self
     }
 
-    #[test]
-    #[ignore]
-    fn dwarf_info_array() {
-        let expected = vec![
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(45))
-                .tag(DwarfTag::DW_TAG_array_type)
-                .type_offset(Offset::new(68))
-                .children(vec![DwarfInfoBuilder::new()
-                    .offset(Offset::new(54))
-                    .tag(DwarfTag::DW_TAG_subrange_type)
-                    .type_offset(Offset::new(61))
-                    .upper_bound(2)
-                    .build()])
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(61))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(8)
-                .name("long unsigned int")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(68))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(4)
-                .name("int")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(75))
-                .tag(DwarfTag::DW_TAG_variable)
-                .name("hoges")
-                .type_offset(Offset::new(45))
-                .location(Location::new(16432))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(97))
-                .tag(DwarfTag::DW_TAG_unimplemented)
-                .name("main")
-                .type_offset(Offset::new(68))
-                .build(),
-        ];
-
-        dwarf_info_intoiterator_test("examples/array", expected);
-    }
-
-    #[test]
-    #[ignore]
-    fn dwarf_info_structure() {
-        let expected = vec![
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(45))
-                .tag(DwarfTag::DW_TAG_structure_type)
-                .name("hoge")
-                .byte_size(8)
-                .children(vec![
-                    DwarfInfoBuilder::new()
-                        .offset(Offset::new(58))
-                        .tag(DwarfTag::DW_TAG_unimplemented)
-                        .name("hoge")
-                        .type_offset(Offset::new(101))
-                        .data_member_location(0)
-                        .build(),
-                    DwarfInfoBuilder::new()
-                        .offset(Offset::new(71))
-                        .tag(DwarfTag::DW_TAG_unimplemented)
-                        .name("fuga")
-                        .type_offset(Offset::new(108))
-                        .data_member_location(4)
-                        .build(),
-                    DwarfInfoBuilder::new()
-                        .offset(Offset::new(84))
-                        .tag(DwarfTag::DW_TAG_unimplemented)
-                        .name("pohe")
-                        .type_offset(Offset::new(115))
-                        .byte_size(4)
-                        .data_member_location(4)
-                        .build(),
-                ])
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(101))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(4)
-                .name("int")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(108))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(1)
-                .name("char")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(115))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(4)
-                .name("unsigned int")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(122))
-                .tag(DwarfTag::DW_TAG_variable)
-                .name("hoge")
-                .type_offset(Offset::new(45))
-                .location(Location::new(16432))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(144))
-                .tag(DwarfTag::DW_TAG_unimplemented)
-                .name("main")
-                .type_offset(Offset::new(101))
-                .build(),
-        ];
-
-        dwarf_info_intoiterator_test("examples/structure", expected);
-    }
-
-    #[test]
-    #[ignore]
-    fn dwarf_info_function_pointer() {
-        let expected = vec![
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(45))
-                .tag(DwarfTag::DW_TAG_subroutine_type)
-                .type_offset(Offset::new(65))
-                .children(vec![
-                    DwarfInfoBuilder::new()
-                        .offset(Offset::new(54))
-                        .tag(DwarfTag::DW_TAG_formal_parameter)
-                        .type_offset(Offset::new(65))
-                        .build(),
-                    DwarfInfoBuilder::new()
-                        .offset(Offset::new(59))
-                        .tag(DwarfTag::DW_TAG_formal_parameter)
-                        .type_offset(Offset::new(72))
-                        .build(),
-                ])
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(65))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(4)
-                .name("int")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(72))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(1)
-                .name("char")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(79))
-                .tag(DwarfTag::DW_TAG_variable)
-                .name("sub2")
-                .type_offset(Offset::new(101))
-                .location(Location::new(16424))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(101))
-                .tag(DwarfTag::DW_TAG_pointer_type)
-                .byte_size(8)
-                .type_offset(Offset::new(45))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(107))
-                .tag(DwarfTag::DW_TAG_unimplemented)
-                .name("main")
-                .type_offset(Offset::new(65))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(137))
-                .tag(DwarfTag::DW_TAG_unimplemented)
-                .name("sub1")
-                .type_offset(Offset::new(65))
-                .children(vec![
-                    DwarfInfoBuilder::new()
-                        .offset(Offset::new(167))
-                        .tag(DwarfTag::DW_TAG_formal_parameter)
-                        .name("i")
-                        .type_offset(Offset::new(65))
-                        .build(),
-                    DwarfInfoBuilder::new()
-                        .offset(Offset::new(179))
-                        .tag(DwarfTag::DW_TAG_formal_parameter)
-                        .name("c")
-                        .type_offset(Offset::new(72))
-                        .build(),
-                ])
-                .build(),
-        ];
-
-        dwarf_info_intoiterator_test("examples/function-pointer", expected);
-    }
-
-    #[test]
-    #[ignore]
-    fn dwarf_info_complex_structure() {
-        let expected = vec![
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(45))
-                .tag(DwarfTag::DW_TAG_structure_type)
-                .name("student")
-                .byte_size(16)
-                .children(vec![DwarfInfoBuilder::new()
-                    .offset(Offset::new(58))
-                    .tag(DwarfTag::DW_TAG_unimplemented)
-                    .name("name")
-                    .type_offset(Offset::new(72))
-                    .data_member_location(0)
-                    .build()])
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(72))
-                .tag(DwarfTag::DW_TAG_array_type)
-                .type_offset(Offset::new(95))
-                .children(vec![DwarfInfoBuilder::new()
-                    .offset(Offset::new(81))
-                    .tag(DwarfTag::DW_TAG_subrange_type)
-                    .type_offset(Offset::new(88))
-                    .upper_bound(15)
-                    .build()])
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(88))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(8)
-                .name("long unsigned int")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(95))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(1)
-                .name("char")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(102))
-                .tag(DwarfTag::DW_TAG_structure_type)
-                .name("hoge")
-                .byte_size(32)
-                .children(vec![
-                    DwarfInfoBuilder::new()
-                        .offset(Offset::new(115))
-                        .tag(DwarfTag::DW_TAG_unimplemented)
-                        .name("hoge")
-                        .type_offset(Offset::new(155))
-                        .data_member_location(0)
-                        .build(),
-                    DwarfInfoBuilder::new()
-                        .offset(Offset::new(128))
-                        .tag(DwarfTag::DW_TAG_unimplemented)
-                        .name("array")
-                        .type_offset(Offset::new(168))
-                        .data_member_location(8)
-                        .build(),
-                    DwarfInfoBuilder::new()
-                        .offset(Offset::new(141))
-                        .tag(DwarfTag::DW_TAG_unimplemented)
-                        .name("student")
-                        .type_offset(Offset::new(45))
-                        .data_member_location(16)
-                        .build(),
-                ])
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(155))
-                .tag(DwarfTag::DW_TAG_pointer_type)
-                .byte_size(8)
-                .type_offset(Offset::new(161))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(161))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(4)
-                .name("int")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(168))
-                .tag(DwarfTag::DW_TAG_array_type)
-                .type_offset(Offset::new(161))
-                .children(vec![DwarfInfoBuilder::new()
-                    .offset(Offset::new(177))
-                    .tag(DwarfTag::DW_TAG_subrange_type)
-                    .type_offset(Offset::new(88))
-                    .upper_bound(1)
-                    .build()])
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(184))
-                .tag(DwarfTag::DW_TAG_array_type)
-                .type_offset(Offset::new(102))
-                .children(vec![DwarfInfoBuilder::new()
-                    .offset(Offset::new(193))
-                    .tag(DwarfTag::DW_TAG_subrange_type)
-                    .type_offset(Offset::new(88))
-                    .upper_bound(2)
-                    .build()])
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(200))
-                .tag(DwarfTag::DW_TAG_variable)
-                .name("hoge")
-                .type_offset(Offset::new(184))
-                .location(Location::new(16480))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(222))
-                .tag(DwarfTag::DW_TAG_unimplemented)
-                .name("main")
-                .type_offset(Offset::new(161))
-                .build(),
-        ];
-
-        dwarf_info_intoiterator_test("examples/complex-structure", expected);
-    }
-
-    #[test]
-    #[ignore]
-    fn dwarf_info_many_compilation_units() {
-        let expected = vec![
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(45))
-                .tag(DwarfTag::DW_TAG_variable)
-                .name("c")
-                .type_offset(Offset::new(65))
-                .location(Location::new(16424))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(65))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(4)
-                .name("int")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(72))
-                .tag(DwarfTag::DW_TAG_unimplemented)
-                .name("main")
-                .type_offset(Offset::new(65))
-                .children(vec![DwarfInfoBuilder::new()
-                    .offset(Offset::new(106))
-                    .tag(DwarfTag::DW_TAG_unimplemented)
-                    .build()])
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(120))
-                .tag(DwarfTag::DW_TAG_unimplemented)
-                .name("sub1")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(182))
-                .tag(DwarfTag::DW_TAG_variable)
-                .name("c")
-                .type_offset(Offset::new(202))
-                .location(Location::new(16424))
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(202))
-                .tag(DwarfTag::DW_TAG_base_type)
-                .byte_size(4)
-                .name("int")
-                .build(),
-            DwarfInfoBuilder::new()
-                .offset(Offset::new(209))
-                .tag(DwarfTag::DW_TAG_unimplemented)
-                .name("sub1")
-                .type_offset(Offset::new(202))
-                .children(vec![DwarfInfoBuilder::new()
-                    .offset(Offset::new(239))
-                    .tag(DwarfTag::DW_TAG_formal_parameter)
-                    .name("i")
-                    .type_offset(Offset::new(202))
-                    .build()])
-                .build(),
-        ];
-
-        dwarf_info_intoiterator_test("examples/many-compilation-units", expected);
+    pub fn children(mut self, children: Vec<DwarfInfo>) -> Self {
+        self.children = children;
+        self
     }
 }
