+++ src/main.rs
@@ -1,13 +1,8 @@
-mod domain;
-mod infrastructure;
-mod library;
-mod usecase;
-
 use env_logger::Env;
 use std::env;
 
-use infrastructure::fromelf::stdout::FromElfStdOut;
-use usecase::dump_global_variables::DumpGlobalVariablesUsecase;
+use troll::infrastructure::fromelf::stdout::FromElfStdOut;
+use troll::usecase::dump_global_variables::DumpGlobalVariablesUsecase;
 
 fn main() {
     env_logger::from_env(Env::default().default_filter_or("warn")).init();
